For the 'Lot-4' Release:
Bug fixes, plus a substantial refactoring (the C# implementation of) the
classes ESBehavior and ESMethod, and also of the AbstractSyntaxTreeNode
classes. The refactoring was for the purpose of supporting the implementation
of Traits.

Other than the bug fixes, the only change visible to those not directly
dealing with the C# code is that instances of Behavior no longer support
the specification of instance variables for their instances. That makes
Behavior instances more like those in Squeak, Pharo or VisualWorks.